requirements.txt
langchain
langchain_community
streamlit
python-dotenv
langchain_groq
pypdf
langchain_core
langserve
fastapi
uvicorn
sse_starlette
pypdf
chromadb
faiss-cpu
groq
cassio
wikipedia
arxiv
langchainhub
sentence_transformers
PyPDF2
langchain-objectbox

